Working allocator and such
The probelem broke down into a few different components.

The first was that in the mapper it was using pages directely mapped to
the kernel space. I replaced this with a scratch allocator, that uses
one table.

The second problem was in the elf parsing, I was not actually finding
the physical regions that corresponed to the elf sections. As such when
I fixed that it ended up working. I'm still not positive about how I'm
supposed to find the physical location... but that's okay.

I then upgraded the mapper to free the page table pages allocated for
pages.